+++ fuint-repository/src/main/java/com/fuint/repository/mapper/TAccountDutyMapper.java
@@ -13,7 +13,7 @@ import java.util.List;
  */
 public interface TAccountDutyMapper extends BaseMapper<TAccountDuty> {
 
-   List<Long> getDutyIdsByAccountId(long accountId);
+   List<Integer> getDutyIdsByAccountId(Integer accountId);
 
    void deleteDutiesByAccountId(long accountId);
